Refactor map into separate component
